Responsive work
Changed a few html items around to make it more responsive. At this
stage I have a nav menu for small screen and a nav menu for large
screen. On the mobile- small screen the link from the nav menu will
take you to a full page. But from the browser- large screen, the page
will toggleClass when clicked on and display NEXT to the browser menu